♻️ refactor(server): split api.rs into domain-focused endpoint modules (#92)
## Why (User Story)
As a developer, I want API endpoints organized by domain so that health,
graph, compute, connections, agents, Athena, and query logic are
independently maintainable.

## What Changed
- Split 2521-line api.rs into 8-file module:
  - mod.rs: Shared types, loading guard, flat re-exports
  - health.rs: Health check, loading, stats, catalog, queue status
  - graph.rs: Graph nodes, node detail, force layout
  - compute.rs: PageRank, communities, degrees, patterns, anomalies
  - connections.rs: Database, queue, Athena connection CRUD
  - agents.rs: Agent list/execute/chat, team execute, sessions
  - athena_query.rs: SSE query streaming, Parquet export, schemas
  - query.rs: Natural language query endpoint

Related to #93
